requirements.txt
uvicorn
asyncpg
fastapi
SQLAlchemy
strawberry-graphql[debug-server]
# surrealdb
# bq_schema
# google-cloud-bigquery
pytz